+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## UNRELEASED
 
+### Features
+
+* [#1665](https://github.com/crypto-org-chain/cronos/pull/1665) Support register for payee and counterpartyPayee in relayer precompile.
+
 ### Improvements
 
+* [#1664](https://github.com/crypto-org-chain/cronos/pull/1664) Update cometbft to 0.38.13.
 * [#]() Add testnet benchmark command.
 
 *Oct 24, 2024*